ResidualBonusSetting CRUD

+++ MLM/database/migrations/2021_30_08_122550_create_residual_bonus_settings_table.php
@@ -19,7 +19,7 @@ class CreateResidualBonusSettingsTable extends Migration
             $table->unsignedBigInteger('rank');
 
             $table->integer('level');
-            $table->integer('percentage');
+            $table->double('percentage');
 
             $table->unique(['rank', 'level']);
